Allow vulkan interop to work with EGL (part 2)
In the GfxstreamBackend EGL mode, swiftshader was mistakenly set as
the vk icd env variable. This still means we are using host vk, but the
meaning of that icd env variable is also to turn off interop. But we
have native interop available in NV EGL anyway, so turn it back on!

Bug: 171711491
Change-Id: I0fac75421efaca766307fe44a05b0ccea027df33