Enable tracing.
This CL enables tracing similar to the tracing implementation
in the deprecated build. As part of this change, pieces of perfetto
needed to be pulled in. Instead of properly supporting protobuf
generation in cmake, we instead land the generated headers.

This CL also does not offer support for the full SDK build.

Bug: 172597927
Test: Build gfxstream_backend.
Change-Id: I011b428d55135b4be92557d905883119c990b65d