vulkan-cereal: external memory ASG
If kFeature_ExternalBlob is enabled, the ASG ring buffer will have
to be allocated via exportable external memory.

When kFeature_ExternalBlob is disabed, we use malloc'ed memory.

This feature goes in hand with upcoming ASG refactorings.  The
new ASG refactorings take effect in the guest requires a blobID
of zero.  That indicates an ASG blob.

Since kInvalidHostmemId is 0, this case will never be hit by
the current protocol but only the future ASG protocol.

Some changes are the base SharedMemory are needed.  For example,
shm_open / shm_unlink aren't available with Bionic, and there
does seem to be a Android build of vulkan cereal.  But no Android
on Android of vulkan-cereal use exists currently, so we just fail
the allocation there.  One can use memfd_create(..), but that's
not available on all Posix systems, just Android/Linux.

BUG=235485545
TEST=compile

Change-Id: I7b66c17c951d4f49c143bc5b7897ebfe434af036