aemu: stop requiring lz4, Googletest, and perfetto-tracing-only
- Test for LZ4_VERSION_RELEASE.  That'll allow one less dependency
for those who don't need it. Nobody seems to use
CompressingStream/DecompressingStream in vulkan-cereal yet?

- Modify third-party/CMakeLists.txt for lz4, Googletest
and perfetto-tracing-only.  These should be optional dependencies
for those who want it.  In addition, we shouldn't depend on
vulkan-cereal to provide the dependencies, since vulkan-cereal
depends on hardware/google/aemu.

- This adds the optional USE_PERFETTO_TRACING flag, which can be
added if the "perfetto-tracing-only" library has been built.

   * In the next few months, perfetto itself should provide
     a C api (https://bit.ly/perfetto-c)(buganizer: 184929776).

BUG=239639751
TEST=compile

Change-Id: I0f23ef32557efd038df153429b78ae6d61aee62b